Create README.md
# BlackJack
Simple Black jack game in Python PyGame

I made this over Christmas break.
It was really just a way for me to learn python and use PyGame a bit.
Did not take that long and the functinality is not bad. 
I wish I would of structured it a bit better.
Hopefully in the future I can remake it and have better visuals as well as a real betting system.